+++ impeller/entity/shaders/linear_to_srgb_filter.vert
@@ -2,12 +2,14 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+#include <impeller/types.glsl>
+
 uniform FrameInfo {
   mat4 mvp;
 } frame_info;
 
-in vec2 position;
-out vec2 v_position;
+in f16vec2 position;
+out f16vec2 v_position;
 
 void main() {
   v_position = position;
