+++ impeller/entity/shaders/texture_fill.frag
@@ -2,22 +2,23 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+#include <impeller/types.glsl>
 #include <impeller/texture.glsl>
 
 uniform sampler2D texture_sampler;
 
 uniform FragInfo {
-  float texture_sampler_y_coord_scale;
-  float alpha;
+  float16_t texture_sampler_y_coord_scale;
+  float16_t alpha;
 }
 frag_info;
 
-in vec2 v_texture_coords;
+in f16vec2 v_texture_coords;
 
 out vec4 frag_color;
 
 void main() {
-  vec4 sampled = IPSample(texture_sampler, v_texture_coords,
+  f16vec4 sampled = IPSample(texture_sampler, v_texture_coords,
                           frag_info.texture_sampler_y_coord_scale);
   frag_color = sampled * frag_info.alpha;
 }
